chore: remove old charts implementation (#1595)
* chore: remove old chart from gql layer

* test: change charts bats tests

* chore: remove old charts module at app level

* chore: remove old charts code from module
