--- bats/admin-gql/new-chart-of-accounts.gql
@@ -1,6 +1,0 @@
-query NewChartOfAccounts {
-  newChartOfAccounts {
-    chartId
-    name
-  }
-}
