+++ core/chart-of-accounts/src/chart_of_accounts/error.rs
@@ -1,5 +1,7 @@
 use thiserror::Error;
 
+use crate::AccountCode;
+
 #[derive(Error, Debug)]
 pub enum ChartError {
     #[error("ChartError - Sqlx: {0}")]
@@ -8,14 +10,8 @@ pub enum ChartError {
     EsEntityError(es_entity::EsEntityError),
     #[error("ChartError - CursorDestructureError: {0}")]
     CursorDestructureError(#[from] es_entity::CursorDestructureError),
-    #[error("ChartError - ChartPathError: '{0}'")]
-    ChartPathError(#[from] crate::path::error::ChartPathError),
-    #[error("ChartError - ControlAccountAlreadyRegistered: '{0}'")]
-    ControlAccountAlreadyRegistered(String),
-    #[error("ChartError - ControlSubAccountAlreadyRegistered: '{0}'")]
-    ControlSubAccountAlreadyRegistered(String),
-    #[error("ChartError - CodeDoesNotExistInChart: '{0}'")]
-    CodeDoesNotExistInChart(String),
+    #[error("ChartError - CodeNotFoundInChart: {0}")]
+    CodeNotFoundInChart(AccountCode),
 }
 
 es_entity::from_es_entity_error!(ChartError);
