+++ core/chart-of-accounts/src/chart_of_accounts/mod.rs
@@ -1,8 +1,10 @@
+mod csv;
 mod entity;
 pub mod error;
 mod repo;
 pub mod tree;
 
+pub(super) use csv::{CsvParseError, CsvParser};
 pub use entity::Chart;
 pub(super) use entity::*;
 pub(super) use repo::*;
