+++ core/chart-of-accounts/src/chart_of_accounts/repo.rs
@@ -2,16 +2,17 @@ use sqlx::PgPool;
 
 use es_entity::*;
 
-use crate::primitives::ChartId;
+use crate::ChartId;
 
 use super::entity::*;
+use super::error::ChartError;
 
 #[derive(EsRepo, Clone)]
 #[es_repo(
     entity = "Chart",
     err = "ChartError",
     columns(reference(ty = "String")),
-    tbl_prefix = "core"
+    tbl_prefix = "core_alt"
 )]
 pub struct ChartRepo {
     pool: PgPool,
