+++ core/chart-of-accounts/src/error.rs
@@ -1,6 +1,6 @@
 use thiserror::Error;
 
-use crate::new::AccountCode;
+pub use crate::chart_of_accounts::error::ChartError;
 
 #[derive(Error, Debug)]
 pub enum CoreChartOfAccountsError {
@@ -9,9 +9,7 @@ pub enum CoreChartOfAccountsError {
     #[error("CoreChartOfAccountsError - AuthorizationError: {0}")]
     AuthorizationError(#[from] authz::error::AuthorizationError),
     #[error("CoreChartOfAccountsError - ChartError: {0}")]
-    ChartError(#[from] crate::chart_of_accounts::error::ChartError),
-    #[error("CoreChartOfAccountsError - AltChartError: {0}")]
-    AltChartError(#[from] crate::new::error::AltChartError),
+    ChartError(#[from] ChartError),
     #[error("CoreChartOfAccountsError - AuditError: {0}")]
     AuditError(#[from] audit::error::AuditError),
     #[error("CoreChartOfAccountsError - CalaLedgerError: {0}")]
@@ -21,7 +19,5 @@ pub enum CoreChartOfAccountsError {
     #[error("CoreChartOfAccountsError - CalaAccountSetError: {0}")]
     CalaAccountSet(#[from] cala_ledger::account_set::error::AccountSetError),
     #[error("CoreChartOfAccountsError - CsvParseError: {0}")]
-    CsvParse(#[from] crate::new::CsvParseError),
-    #[error("CoreChartOfAccountsError - AccountNotFoundInChart: {0}")]
-    AccountNotFoundInChart(AccountCode),
+    CsvParse(#[from] crate::CsvParseError),
 }
