+++ core/credit/src/error.rs
@@ -26,10 +26,10 @@ pub enum CoreCreditError {
     PriceError(#[from] core_price::error::PriceError),
     #[error("CoreCreditError - GovernanceError: {0}")]
     GovernanceError(#[from] governance::error::GovernanceError),
-    #[error("CoreDepositError - CoreChartOfAccountsError: {0}")]
+    #[error("CoreCreditError - CoreChartOfAccountsError: {0}")]
     CoreChartOfAccountsError(#[from] chart_of_accounts::error::CoreChartOfAccountsError),
-    #[error("CoreDepositError - CoreChartOfAccountsError: {0}")]
-    AltCoreChartOfAccountsError(#[from] chart_of_accounts::new::error::AltChartError),
+    #[error("CoreCreditError - CoreChartOfAccountsError: {0}")]
+    AltCoreChartOfAccountsError(#[from] chart_of_accounts::error::ChartError),
     #[error("CoreCreditError - JobError: {0}")]
     JobError(#[from] job::error::JobError),
     #[error("CoreCreditError - CustomerMismatchForCreditFacility")]
