+++ lana/admin-server/src/graphql/chart_of_accounts.rs
@@ -8,6 +8,7 @@ use lana_app::chart_of_accounts::Chart as DomainChart;
 #[graphql(complex)]
 pub struct ChartOfAccounts {
     id: ID,
+    chart_id: UUID,
     name: String,
 
     #[graphql(skip)]
@@ -18,6 +19,7 @@ impl From<DomainChart> for ChartOfAccounts {
     fn from(chart: DomainChart) -> Self {
         ChartOfAccounts {
             id: chart.id.to_global_id(),
+            chart_id: UUID::from(chart.id),
             name: chart.name.to_string(),
 
             entity: Arc::new(chart),
@@ -27,128 +29,40 @@ impl From<DomainChart> for ChartOfAccounts {
 
 #[ComplexObject]
 impl ChartOfAccounts {
-    async fn categories(&self) -> ChartCategories {
-        let tree = self.entity.chart();
-        ChartCategories {
-            assets: ChartCategory {
-                name: tree.assets.name,
-                account_code: tree.assets.encoded_path,
-                control_accounts: tree
-                    .assets
-                    .children
-                    .into_iter()
-                    .map(ChartControlAccount::from)
-                    .collect(),
-            },
-            liabilities: ChartCategory {
-                name: tree.liabilities.name,
-                account_code: tree.liabilities.encoded_path,
-                control_accounts: tree
-                    .liabilities
-                    .children
-                    .into_iter()
-                    .map(ChartControlAccount::from)
-                    .collect(),
-            },
-            equity: ChartCategory {
-                name: tree.equity.name,
-                account_code: tree.equity.encoded_path,
-                control_accounts: tree
-                    .equity
-                    .children
-                    .into_iter()
-                    .map(ChartControlAccount::from)
-                    .collect(),
-            },
-            revenues: ChartCategory {
-                name: tree.revenues.name,
-                account_code: tree.revenues.encoded_path,
-                control_accounts: tree
-                    .revenues
-                    .children
-                    .into_iter()
-                    .map(ChartControlAccount::from)
-                    .collect(),
-            },
-            expenses: ChartCategory {
-                name: tree.expenses.name,
-                account_code: tree.expenses.encoded_path,
-                control_accounts: tree
-                    .expenses
-                    .children
-                    .into_iter()
-                    .map(ChartControlAccount::from)
-                    .collect(),
-            },
-        }
+    async fn children(&self) -> Vec<ChartNode> {
+        self.entity
+            .chart()
+            .children
+            .into_iter()
+            .map(ChartNode::from)
+            .collect()
     }
 }
 
 #[derive(SimpleObject)]
-pub struct ChartCategories {
-    assets: ChartCategory,
-    liabilities: ChartCategory,
-    equity: ChartCategory,
-    revenues: ChartCategory,
-    expenses: ChartCategory,
-}
-
-#[derive(SimpleObject)]
-pub struct ChartCategory {
+pub struct ChartNode {
     name: String,
     account_code: String,
-    control_accounts: Vec<ChartControlAccount>,
+    children: Vec<ChartNode>,
 }
 
-#[derive(SimpleObject)]
-pub struct ChartControlAccountDetails {
-    name: String,
-    account_code: String,
-}
-
-impl From<lana_app::chart_of_accounts::ControlAccountDetails> for ChartControlAccountDetails {
-    fn from(details: lana_app::chart_of_accounts::ControlAccountDetails) -> Self {
+impl From<lana_app::chart_of_accounts::tree::TreeNode> for ChartNode {
+    fn from(node: lana_app::chart_of_accounts::tree::TreeNode) -> Self {
         Self {
-            name: details.name,
-            account_code: details.path.to_string(),
+            name: node.name.to_string(),
+            account_code: node.code.to_string(),
+            children: node.children.into_iter().map(ChartNode::from).collect(),
         }
     }
+}
+
+#[derive(InputObject)]
+pub struct ChartOfAccountsCsvImportInput {
+    pub chart_id: UUID,
+    pub file: Upload,
 }
 
 #[derive(SimpleObject)]
-pub struct ChartControlAccount {
-    name: String,
-    account_code: String,
-    control_sub_accounts: Vec<ChartControlSubAccount>,
-}
-
-impl From<lana_app::chart_of_accounts::tree::ChartTreeControlAccount> for ChartControlAccount {
-    fn from(tree: lana_app::chart_of_accounts::tree::ChartTreeControlAccount) -> Self {
-        ChartControlAccount {
-            name: tree.name,
-            account_code: tree.encoded_path,
-            control_sub_accounts: tree
-                .children
-                .into_iter()
-                .map(ChartControlSubAccount::from)
-                .collect(),
-        }
-    }
-}
-
-#[derive(SimpleObject)]
-pub struct ChartControlSubAccount {
-    name: String,
-    account_code: String,
-}
-
-impl From<lana_app::chart_of_accounts::tree::ChartTreeControlSubAccount>
-    for ChartControlSubAccount
-{
-    fn from(tree: lana_app::chart_of_accounts::tree::ChartTreeControlSubAccount) -> Self {
-        ChartControlSubAccount {
-            name: tree.name,
-            account_code: tree.encoded_path,
-        }
-    }
+pub struct ChartOfAccountsCsvImportPayload {
+    pub success: bool,
 }
