+++ lana/app/src/accounting_init/constants/mod.rs
@@ -3,8 +3,6 @@ pub(super) const LANA_JOURNAL_CODE: &str = "LANA_BANK_JOURNAL";
 pub const CHART_REF: &str = "primary-chart";
 pub(super) const CHART_NAME: &str = "Chart of Accounts";
 
-pub const OBS_CHART_REF: &str = "off-balance-sheet-chart";
-
 pub const TRIAL_BALANCE_STATEMENT_NAME: &str = "Trial Balance";
 pub const OBS_TRIAL_BALANCE_STATEMENT_NAME: &str = "Off-Balance-Sheet Trial Balance";
